@fraguada @pearswj
Here is the 4th assignment. I had no problem to make it at this time since the Appserver is automated to do it. However I think that I can't make this kind of server system . I understood about 35% of the class how it works so that I'll need a support session tomorrow...
Thank you so much for your supports.